ch09-templates/python/requirements.txt
lxml
jinja2
pyyaml